# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UpFlow - Regulação de Leitos API",
        "description": "API REST para gerenciamento e controle de leitos hospitalares desenvolvida em Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Ambiente Local"
        }
    ],
    "paths": {
...
        "/api/leitos": {
            "get": {
                "tags": [
                    "Leitos"
                ],
                "summary": "Lista todos os leitos",
                "operationId": "1c885723bd60762c698dfcff3917aae3",
                "responses": {
                    "200": {
                        "description": "Lista de leitos retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Leitos"
                ],
                "summary": "Cadastra um novo leito",
                "operationId": "11e679a048acd3bd4b6ddf24d2d92add",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero",
                                    "tipo_leito_id",
                                    "ativo"
                                ],
                                "properties": {
                                    "numero": {
                                        "type": "string",
                                        "example": "UTI-03"
                                    },
                                    "tipo_leito_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leito cadastrado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/leitos/{leito}": {
            "get": {
                "tags": [
                    "Leitos"
                ],
                "summary": "Exibe um leito específico",
                "operationId": "625053572e8b7dec97c321af0f1a0e54",
                "parameters": [
                    {
                        "name": "leito",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leito retornado com sucesso"
                    },
                    "404": {
                        "description": "Leito não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Leitos"
                ],
                "summary": "Atualiza um leito",
                "operationId": "d2c219eee89d75f1b638508a53023586",
                "parameters": [
                    {
                        "name": "leito",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero": {
                                        "type": "string",
                                        "example": "UTI-03"
                                    },
                                    "tipo_leito_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Leito atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leitos"
                ],
                "summary": "Remove um leito",
                "operationId": "1d41c009cb2dfd49a842a531c5d87e99",
                "parameters": [
                    {
                        "name": "leito",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Leito removido com sucesso"
                    },
                    "404": {
                        "description": "Leito não encontrado"
                    }
                }
            }
        },
        "/api/leitos/{leito}/status": {
            "get": {
                "tags": [
                    "Leitos"
                ],
                "summary": "Verifica o status de ocupação de um leito",
                "description": "Retorna se o leito está LIVRE ou OCUPADO. Caso esteja ocupado, retorna o paciente atual.",
                "operationId": "9865b5f8f673118a226caf006fa00558",
                "parameters": [
                    {
                        "name": "leito",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status do leito retornado com sucesso"
                    },
                    "404": {
                        "description": "Leito não encontrado"
                    }
                }
            }
        },
...
    },
...
}